requirements.txt
fastapi==0.134.0
uvicorn==0.41.0
pydantic==2.12.5
numpy==2.4.2
pandas==3.0.1
scikit-learn==1.8.0
joblib==1.5.3
pytesseract==0.3.13
pillow==12.1.1
opencv-python==4.13.0.92
python-multipart==0.0.22
PyMuPDF==1.27.2.2
python-docx==1.2.0

# Added missing dependencies required for Core Bot functionality
google-api-python-client==2.193.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
beautifulsoup4==4.14.3
reportlab==4.4.10
requests==2.32.5
python-whois==0.9.6